requirements.txt
py-cord[voice]
asyncio
PyNaCl
python-dotenv
yt-dlp
spotipy